feat: tonnage converted to double

+++ database/migrations/2023_04_24_110813_create_customers_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
             $table->integer("tb")->default(0)->nullable();
             $table->integer("tw")->default(0)->nullable();
             $table->integer("thr")->default(0)->nullable();
-            $table->integer("tonnage")->default(0)->nullable();
+            $table->double("tonnage")->default(0)->nullable();
             $table->integer("cashback_approved")->default(0)->nullable();
             $table->timestamps();
         });
