+++ database/migrations/2023_04_24_192600_create_rit_branches_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('rit_branches', function (Blueprint $table) {
             $table->id();
             $table->string("name");
-            $table->integer("sent_tonnage");
+            $table->double("sent_tonnage");
             $table->integer("income")->nullable();
             $table->date("delivery_date");
             $table->timestamps();
