+++ database/migrations/2023_04_24_193252_create_incomes_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('incomes', function (Blueprint $table) {
             $table->id();
             $table->integer("amount");
-            $table->integer("tonnage");
+            $table->double("tonnage");
             $table->string("type");
             $table->timestamps();
         });
